Clean up schemacode a bit.
